Group related tests into combined files
