--- tests/BraceComment.cs
@@ -1,9 +1,0 @@
-namespace Demo {
-    public partial class Foo {
-        public int Value() {
-            int result;
-            result = 123;
-            return result;
-        }
-    }
-}
