--- tests/CaseElse.cs
@@ -1,27 +1,0 @@
-using System;
-
-namespace N {
-    public partial class TTest {
-        public void Foo(int x) {
-            switch (x)
-            {
-                case 1: Console.WriteLine("one"); break;
-                default:
-                {
-                    Console.WriteLine("other");
-                break;
-                }
-            }
-        }
-        public void Empty(int x) {
-            switch (x)
-            {
-                case 1: Console.WriteLine("one"); break;
-                default:
-                {
-                break;
-                }
-            }
-        }
-    }
-}
