--- tests/Commented.cs
@@ -1,7 +1,0 @@
-namespace Demo {
-    public partial class Commented {
-        public void SayHello() {
-            System.Console.WriteLine("Hello");
-        }
-    }
-}
