--- tests/IfExpr.cs
@@ -1,11 +1,0 @@
-namespace Demo {
-    public partial class Ternary {
-        public static int Select(bool flag) {
-            int result;
-            int val;
-            val = flag ? 1 : 0;
-            result = val;
-            return result;
-        }
-    }
-}
