--- tests/IfOr.cs
@@ -1,13 +1,0 @@
-namespace Demo {
-    public partial class Foo {
-        // TODO: field ehVazio: bool -> declare a field
-        public bool ehVazio;
-        // TODO: field indUltimaPos: int -> declare a field
-        public int indUltimaPos;
-        public bool Check(int frowInd) {
-            bool result;
-            if (this.ehVazio || frowInd > this.indUltimaPos || frowInd == -1) result = true; else result = false;
-            return result;
-        }
-    }
-}
