--- tests/ParenComment.pas
@@ -1,18 +1,0 @@
-namespace Demo;
-
-type
-  Foo = class
-  public
-    method Value: Integer;
-  end;
-
-implementation
-
-method Foo.Value: Integer;
-begin
-  (* comment with *** inside
-     multiple lines *)
-  result := 123;
-end;
-
-end.
