--- tests/TryExceptEmpty.cs
@@ -1,14 +1,0 @@
-using System;
-
-namespace Demo {
-    public partial class TryExceptEmptyExample {
-        public static void DoNothing() {
-            try
-            {
-                Console.WriteLine("A");
-            }
-            catch (Exception)
-            {}
-        }
-    }
-}
