--- tests/TryExceptEmptySemi.cs
@@ -1,24 +1,0 @@
-using System;
-
-namespace Demo {
-    public partial class TryExceptEmptySemiExample {
-        public static void DoNothing() {
-            try
-            {
-                Console.WriteLine("A");
-            }
-            catch (Exception)
-            {}
-        }
-        public static void DoHandle() {
-            try
-            {
-                Console.WriteLine("B");
-            }
-            catch (Exception)
-            {
-                Console.WriteLine("Error");
-            }
-        }
-    }
-}
