--- tests/TryExceptOn.cs
@@ -1,16 +1,0 @@
-using System;
-
-namespace Demo {
-    public partial class TryExceptExample {
-        public static void DoStuff() {
-            try
-            {
-                Console.WriteLine("A");
-            }
-            catch (Exception E)
-            {
-                Console.WriteLine(E.Message);
-            }
-        }
-    }
-}
