--- tests/TryExceptOnType.cs
@@ -1,16 +1,0 @@
-using System;
-
-namespace Demo {
-    public partial class TryExceptOnType {
-        public static void DoStuff() {
-            try
-            {
-                Console.WriteLine("A");
-            }
-            catch (Exception)
-            {
-                Console.WriteLine("Error");
-            }
-        }
-    }
-}
